+++ design-system/icon-lab/core/chart-area.d.ts
@@ -1,0 +1,13 @@
+/**
+ * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
+ * @codegen <<SignedSource::32ed27d357dc456877fd71d65ba46491>>
+ * @codegenCommand yarn build:icon-glyphs
+ */
+
+import type { NewCoreIconProps } from '@atlaskit/icon/base-new';
+
+declare const ChartAreaIcon: {
+  (props: NewCoreIconProps): JSX.Element;
+  displayName: string;
+};
+export default ChartAreaIcon;