+++ design-system/icon-lab/core/chart-area.js
@@ -1,6 +1,6 @@
 /**
  * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
- * @codegen <<SignedSource::94c5d2c9bdcdecc62580042a534c4168>>
+ * @codegen <<SignedSource::baef9982ec2dab1b366ea4e66c5e12d1>>
  * @codegenCommand yarn build:icon-glyphs
  */
 "use strict";
@@ -13,16 +13,16 @@ var _react = _interopRequireDefault(require("react"));
 var _baseNew = _interopRequireDefault(require("@atlaskit/icon/base-new"));
 function _interopRequireDefault(e) { return e && e.__esModule ? e : { default: e }; }
 /**
- * Icon: "ChartPie".
+ * Icon: "ChartArea".
  * Category: multi-purpose
- * Location: @atlaskit/icon
- * Usage guidance: Known uses: Pie and segment charts
+ * Location: @atlaskit/icon-lab
+ * Usage guidance: Known uses: Area charts
  * - [Examples](https://atlaskit.atlassian.com/packages/design-system/icon)
  * - [Code](https://atlaskit.atlassian.com/packages/design-system/icon/docs/custom-icons)
  */
-const ChartPieIcon = props => /*#__PURE__*/_react.default.createElement(_baseNew.default, Object.assign({
-  dangerouslySetGlyph: `<path fill="currentcolor" fill-rule="evenodd" d="M7.25 1.543a6.501 6.501 0 1 0 7.207 7.207H8A.75.75 0 0 1 7.25 8zM0 8a8 8 0 1 1 16 0A8 8 0 0 1 0 8" clip-rule="evenodd"/>`
+const ChartAreaIcon = props => /*#__PURE__*/_react.default.createElement(_baseNew.default, Object.assign({
+  dangerouslySetGlyph: `<path fill="currentcolor" d="M1 13V1h1.5v12a.5.5 0 0 0 .5.5h12V15H3a2 2 0 0 1-2-2"/><path fill="currentcolor" d="M14 3v1.5h-3L7.2 7.35a.75.75 0 0 1-.45.15H3.5V6h3l3.8-2.85a.75.75 0 0 1 .45-.15zm-.75 4a.75.75 0 0 1 .75.75v4a.75.75 0 0 1-.75.75h-9a.75.75 0 0 1-.75-.75v-1.5a.75.75 0 0 1 .75-.75h2.79l3.82-2.39.09-.05a.75.75 0 0 1 .3-.06z"/>`
   // eslint-disable-next-line @repo/internal/react/no-unsafe-spread-props
 }, props));
-ChartPieIcon.displayName = 'ChartPieIcon';
-var _default = exports.default = ChartPieIcon;
+ChartAreaIcon.displayName = 'ChartAreaIcon';
+var _default = exports.default = ChartAreaIcon;