+++ design-system/icon-lab/core/chart-bar-line.d.ts
@@ -1,0 +1,13 @@
+/**
+ * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
+ * @codegen <<SignedSource::19fc92d98c8cb67d888b3f6da0a84bac>>
+ * @codegenCommand yarn build:icon-glyphs
+ */
+
+import type { NewCoreIconProps } from '@atlaskit/icon/base-new';
+
+declare const ChartBarLineIcon: {
+  (props: NewCoreIconProps): JSX.Element;
+  displayName: string;
+};
+export default ChartBarLineIcon;