+++ design-system/icon-lab/core/chart-box-plot.js
@@ -1,0 +1,28 @@
+/**
+ * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
+ * @codegen <<SignedSource::6c8c010f1ad18d91852e52d914803559>>
+ * @codegenCommand yarn build:icon-glyphs
+ */
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.default = void 0;
+var _react = _interopRequireDefault(require("react"));
+var _baseNew = _interopRequireDefault(require("@atlaskit/icon/base-new"));
+function _interopRequireDefault(e) { return e && e.__esModule ? e : { default: e }; }
+/**
+ * Icon: "ChartBoxPlot".
+ * Category: multi-purpose
+ * Location: @atlaskit/icon-lab
+ * Usage guidance: Known uses: Box plot charts
+ * - [Examples](https://atlaskit.atlassian.com/packages/design-system/icon)
+ * - [Code](https://atlaskit.atlassian.com/packages/design-system/icon/docs/custom-icons)
+ */
+const ChartBoxPlotIcon = props => /*#__PURE__*/_react.default.createElement(_baseNew.default, Object.assign({
+  dangerouslySetGlyph: `<path fill="currentcolor" d="M1 13V1h1.5v12a.5.5 0 0 0 .5.5h12V15H3a2 2 0 0 1-2-2"/><path fill="currentcolor" d="M7 2v2.5h.75a.75.75 0 0 1 .75.75v4a.75.75 0 0 1-.75.75H7v2H5.5v-2h-.75A.75.75 0 0 1 4 9.25v-4a.75.75 0 0 1 .75-.75h.75V2zM5.5 8.5H7V6H5.5zM13 1v2h.83a.75.75 0 0 1 .67.75v4.5a.75.75 0 0 1-.75.75H13v3h-1.5V9h-.75a.75.75 0 0 1-.75-.75v-4.5a.75.75 0 0 1 .75-.75h.75V1zm-1.5 6.5H13v-3h-1.5z"/>`
+  // eslint-disable-next-line @repo/internal/react/no-unsafe-spread-props
+}, props));
+ChartBoxPlotIcon.displayName = 'ChartBoxPlotIcon';
+var _default = exports.default = ChartBoxPlotIcon;