+++ design-system/icon-lab/core/signpost.d.ts
@@ -1,0 +1,13 @@
+/**
+ * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
+ * @codegen <<SignedSource::d5bd5fd6a108184119212852d428b191>>
+ * @codegenCommand yarn build:icon-glyphs
+ */
+
+import type { NewCoreIconProps } from '@atlaskit/icon/base-new';
+
+declare const SignpostIcon: {
+  (props: NewCoreIconProps): JSX.Element;
+  displayName: string;
+};
+export default SignpostIcon;