+++ design-system/icon-lab/core/signpost.js
@@ -1,6 +1,6 @@
 /**
  * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
- * @codegen <<SignedSource::94c5d2c9bdcdecc62580042a534c4168>>
+ * @codegen <<SignedSource::0a68a8b9adb42339f874d72d1865b4c0>>
  * @codegenCommand yarn build:icon-glyphs
  */
 "use strict";
@@ -13,16 +13,15 @@ var _react = _interopRequireDefault(require("react"));
 var _baseNew = _interopRequireDefault(require("@atlaskit/icon/base-new"));
 function _interopRequireDefault(e) { return e && e.__esModule ? e : { default: e }; }
 /**
- * Icon: "ChartPie".
+ * Icon: "Signpost".
  * Category: multi-purpose
- * Location: @atlaskit/icon
- * Usage guidance: Known uses: Pie and segment charts
+ * Location: @atlaskit/icon-lab
  * - [Examples](https://atlaskit.atlassian.com/packages/design-system/icon)
  * - [Code](https://atlaskit.atlassian.com/packages/design-system/icon/docs/custom-icons)
  */
-const ChartPieIcon = props => /*#__PURE__*/_react.default.createElement(_baseNew.default, Object.assign({
-  dangerouslySetGlyph: `<path fill="currentcolor" fill-rule="evenodd" d="M7.25 1.543a6.501 6.501 0 1 0 7.207 7.207H8A.75.75 0 0 1 7.25 8zM0 8a8 8 0 1 1 16 0A8 8 0 0 1 0 8" clip-rule="evenodd"/>`
+const SignpostIcon = props => /*#__PURE__*/_react.default.createElement(_baseNew.default, Object.assign({
+  dangerouslySetGlyph: `<path fill="currentcolor" d="M3 4.5a.5.5 0 0 0-.5.5v3a.5.5 0 0 0 .5.5h8.792a.5.5 0 0 0 .323-.118L14.338 6.5l-2.223-1.882a.5.5 0 0 0-.323-.118zM7 3V0h1.5v3h3.292a2 2 0 0 1 1.292.474l2.337 1.976c.649.55.649 1.55 0 2.1l-2.337 1.976a2 2 0 0 1-1.292.474H8.5v6H7v-6H3a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2z"/>`
   // eslint-disable-next-line @repo/internal/react/no-unsafe-spread-props
 }, props));
-ChartPieIcon.displayName = 'ChartPieIcon';
-var _default = exports.default = ChartPieIcon;
+SignpostIcon.displayName = 'SignpostIcon';
+var _default = exports.default = SignpostIcon;