+++ design-system/icon-lab/core/terminal.d.ts
@@ -1,0 +1,13 @@
+/**
+ * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
+ * @codegen <<SignedSource::aeee223b51039a34b495fc135878c81e>>
+ * @codegenCommand yarn build:icon-glyphs
+ */
+
+import type { NewCoreIconProps } from '@atlaskit/icon/base-new';
+
+declare const TerminalIcon: {
+  (props: NewCoreIconProps): JSX.Element;
+  displayName: string;
+};
+export default TerminalIcon;