+++ design-system/icon/core/chart-bubble.d.ts
@@ -1,0 +1,13 @@
+/**
+ * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
+ * @codegen <<SignedSource::10cc4b95d4fae99e952f09320d9d14c4>>
+ * @codegenCommand yarn build:icon-glyphs
+ */
+
+import type { NewCoreIconProps } from '@atlaskit/icon/base-new';
+
+declare const ChartBubbleIcon: {
+  (props: NewCoreIconProps): JSX.Element;
+  displayName: string;
+};
+export default ChartBubbleIcon;