+++ design-system/icon/core/chart-trend-down.d.ts
@@ -1,0 +1,13 @@
+/**
+ * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
+ * @codegen <<SignedSource::988616485cfd28d105e9c5ad8da876e9>>
+ * @codegenCommand yarn build:icon-glyphs
+ */
+
+import type { NewCoreIconProps } from '@atlaskit/icon/base-new';
+
+declare const ChartTrendDownIcon: {
+  (props: NewCoreIconProps): JSX.Element;
+  displayName: string;
+};
+export default ChartTrendDownIcon;