+++ design-system/icon/core/chart-trend-up.d.ts
@@ -1,0 +1,13 @@
+/**
+ * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
+ * @codegen <<SignedSource::6b652d6b211dc4e799eadf43e6431d59>>
+ * @codegenCommand yarn build:icon-glyphs
+ */
+
+import type { NewCoreIconProps } from '@atlaskit/icon/base-new';
+
+declare const ChartTrendUpIcon: {
+  (props: NewCoreIconProps): JSX.Element;
+  displayName: string;
+};
+export default ChartTrendUpIcon;