+++ design-system/icon/core/table.d.ts
@@ -1,0 +1,13 @@
+/**
+ * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
+ * @codegen <<SignedSource::bbb9b5a679027ef2c8c367e3e5fd5eed>>
+ * @codegenCommand yarn build:icon-glyphs
+ */
+
+import type { NewCoreIconProps } from '@atlaskit/icon/base-new';
+
+declare const TableIcon: {
+  (props: NewCoreIconProps): JSX.Element;
+  displayName: string;
+};
+export default TableIcon;