+++ design-system/icon/core/table.js
@@ -1,6 +1,6 @@
 /**
  * THIS FILE WAS CREATED VIA CODEGEN DO NOT MODIFY {@see http://go/af-codegen}
- * @codegen <<SignedSource::94c5d2c9bdcdecc62580042a534c4168>>
+ * @codegen <<SignedSource::068663bc7237a4f6d91e480dd5b2499a>>
  * @codegenCommand yarn build:icon-glyphs
  */
 "use strict";
@@ -13,16 +13,16 @@ var _react = _interopRequireDefault(require("react"));
 var _baseNew = _interopRequireDefault(require("@atlaskit/icon/base-new"));
 function _interopRequireDefault(e) { return e && e.__esModule ? e : { default: e }; }
 /**
- * Icon: "ChartPie".
- * Category: multi-purpose
+ * Icon: "Table".
+ * Category: single-purpose
  * Location: @atlaskit/icon
- * Usage guidance: Known uses: Pie and segment charts
+ * Usage guidance: Reserved for tables, tabular data, and spreadsheets.
  * - [Examples](https://atlaskit.atlassian.com/packages/design-system/icon)
  * - [Code](https://atlaskit.atlassian.com/packages/design-system/icon/docs/custom-icons)
  */
-const ChartPieIcon = props => /*#__PURE__*/_react.default.createElement(_baseNew.default, Object.assign({
-  dangerouslySetGlyph: `<path fill="currentcolor" fill-rule="evenodd" d="M7.25 1.543a6.501 6.501 0 1 0 7.207 7.207H8A.75.75 0 0 1 7.25 8zM0 8a8 8 0 1 1 16 0A8 8 0 0 1 0 8" clip-rule="evenodd"/>`
+const TableIcon = props => /*#__PURE__*/_react.default.createElement(_baseNew.default, Object.assign({
+  dangerouslySetGlyph: `<path fill="currentcolor" d="M14.5 8.75h-8v3.75H14a.5.5 0 0 0 .5-.5zM1.5 12a.5.5 0 0 0 .5.5h3V8.75H1.5zm13-8a.5.5 0 0 0-.5-.5H6.5v3.75h8zm-13 3.25H5V3.5H2a.5.5 0 0 0-.5.5zM16 12a2 2 0 0 1-2 2H2a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h12a2 2 0 0 1 2 2z"/>`
   // eslint-disable-next-line @repo/internal/react/no-unsafe-spread-props
 }, props));
-ChartPieIcon.displayName = 'ChartPieIcon';
-var _default = exports.default = ChartPieIcon;
+TableIcon.displayName = 'TableIcon';
+var _default = exports.default = TableIcon;